+++ src/entity.cpp
@@ -12,7 +12,7 @@ Entity::Entity()
 {
 }
 
-void Entity::set_matrix(int idx, glm::mat4 mat)
+void Entity::set_matrix(size_t idx, glm::mat4 mat)
 {
     if (idx >= matrices.size())
         std::cerr << "Entity: invalid index for matrix setting" << std::endl;
@@ -23,13 +23,13 @@ void Entity::draw(Shader shader)
 {
     if (type == EntityType::MODEL)
     {
-        for (int i = 0; i < models.size(); ++i)
+        for (size_t i = 0; i < models.size(); ++i)
             models[i].draw(shader, matrices[i]);
     }
     else if (type == EntityType::MESH)
     {
         glDisable(GL_CULL_FACE);
-        for (int i = 0; i < meshes.size(); ++i)
+        for (size_t i = 0; i < meshes.size(); ++i)
             meshes[i].draw(shader, matrices[i]);
 
         glEnable(GL_CULL_FACE);
